move to separate ts project
