+++ client/cody/CONTRIBUTING.md
@@ -18,6 +18,7 @@
 - `webviews`: source code of the extension UI (webviews),
   build with Vite and rollup.js using the `vite.config.ts` file at directory
   root
+- `integration-test`: code for integration tests
 - `dist`: build outputs from both webpack and vite
 - `resources`: everything in this directory will be move to
   the ./dist directory automatically during build time for easy packaging
