+++ client/cody/integration-test/.eslintrc.js
@@ -1,0 +1,10 @@
+// @ts-check
+
+const baseConfig = require('../.eslintrc')
+module.exports = {
+  extends: '../.eslintrc.js',
+  parserOptions: {
+    ...baseConfig.parserOptions,
+    project: [__dirname + '/tsconfig.json'],
+  },
+}
