+++ client/cody/integration-test/api.test.ts
@@ -2,7 +2,7 @@ import * as assert from 'assert'
 
 import * as vscode from 'vscode'
 
-import { History } from '../completions/history'
+import { History } from '../src/completions/history'
 
 suite('API tests', () => {
     test('Cody registers some commands', async () => {
