+++ client/cody/integration-test/main.ts
@@ -12,19 +12,19 @@ async function main(): Promise<void> {
     // with exit status 0).
     process.env.MOCK_RG_PATH = 'true'
 
-    // When run, this script's filename is `client/cody/out/src/integration-test/runTest.js`, so
+    // When run, this script's filename is `client/cody/out/integration-test/main.js`, so
     // __dirname is derived from that path, not this file's source path.
-    const codyRoot = path.resolve(__dirname, '..', '..', '..')
+    const codyRoot = path.resolve(__dirname, '..', '..')
 
     // The test workspace is not copied to out/ during the TypeScript build, so we need to refer to
     // it in the src/ dir.
-    const testWorkspacePath = path.resolve(codyRoot, 'src', 'integration-test', 'testdata', 'workspace')
+    const testWorkspacePath = path.resolve(codyRoot, 'integration-test', 'testdata', 'workspace')
 
     // The directory containing the extension's package.json, passed to --extensionDevelopmentPath.
     const extensionDevelopmentPath = codyRoot
 
     // The path to the test runner script, passed to --extensionTestsPath.
-    const extensionTestsPath = path.resolve(codyRoot, 'out', 'src', 'integration-test', 'index')
+    const extensionTestsPath = path.resolve(codyRoot, 'out', 'integration-test', 'index')
 
     try {
         // Download VS Code, unzip it, and run the integration test.
