+++ client/cody/integration-test/testdata/workspace/.gitignore
